requirements.txt
opencv-python
pyzbar
numpy
RPi.GPIO
pyserial-asyncio
PyYAML 